+++ src/transfer/transfer.ts
@@ -1,3 +1,4 @@
+import { readFile } from "fs-extra";
 import { Uri } from "vscode";
 import { System } from "../extension/system";
 import { UserConfig, configManager } from "../modules/config";
@@ -7,6 +8,7 @@ import { QuickPickItem } from "../ui/quickpick";
 import statusBar, { Icon } from "../ui/statusbar";
 import { GetUserManager } from "../user/usermanager";
 import { UploadFolderLimited } from "./limited/upload";
+import { UploadFile } from "./upload";
 import path = require("path");
 
 
@@ -25,18 +27,38 @@ export async function TransferFolder(uri: Uri, userConfig: UserConfig) {
         if (!await user.login()) return;
 
         statusBar.updateBar('Transfering', Icon.spinLoading, { duration: -1 });
-        logger.infos("Transfer Folder", path.basename(uri.fsPath) + ": Started");
+        logger.infoplus(configManager.CurrentSystem.name, "Transfer Folder", "->" + system.name + ', ' + path.basename(uri.fsPath) + ": Started");
 
         const response = await UploadFolderLimited(uri.fsPath, userConfig, system);
 
         if (response.aborted) {
             statusBar.updateBar('Cancelled', Icon.success, { duration: 1 });
-            logger.infos("Transfer Folder", path.basename(uri.fsPath) + ": Cancelled");
+            logger.infoplus(configManager.CurrentSystem.name, "Transfer Folder", "->" + system.name + ', ' + path.basename(uri.fsPath) + ": Cancelled");
         }
         else {
             statusBar.updateBar('Transferred', Icon.success, { duration: 1 });
-            logger.infos("Transfer Folder", path.basename(uri.fsPath) + ": Completed");
+            logger.infoplus(configManager.CurrentSystem.name, "Transfer Folder", "->" + system.name + ', ' + path.basename(uri.fsPath) + ": Completed");
         }
+    }
+
+}
+
+export async function TransferFile(uri: Uri, userConfig: UserConfig) {
+    let picks: QuickPickItem<System>[] = [];
+    for (const system of userConfig.systems) {
+        if (configManager.CurrentSystem != system) {
+            picks.push({ label: system.name, description: system.host + ':' + system.port, object: system })
+        }
+    }
+
+    const quickResponse: QuickPickItem<System> = await ShowQuickPick(picks, { title: 'Pick System' });
+    if (quickResponse) {
+        const system = quickResponse.object;
+        const user = GetUserManager(system, true);
+        if (!await user.login()) return;
+
+        const content = (await readFile(uri.fsPath)).toString('utf-8');
+        const response = await UploadFile(uri, content, userConfig, system);
     }
 
 }